src/test/javascript/spec/app/entities/fincasMs/ubicacion/ubicacion.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { UbicacionService } from 'app/entities/fincasMs/ubicacion/ubicacion.service';
import { IUbicacion, Ubicacion } from 'app/shared/model/fincasMs/ubicacion.model';
import { Departamento } from 'app/shared/model/enumerations/departamento.model';

describe('Service Tests', () => {
  describe('Ubicacion Service', () => {
    let injector: TestBed;
    let service: UbicacionService;
    let httpMock: HttpTestingController;
    let elemDefault: IUbicacion;
    let expectedResult: IUbicacion | IUbicacion[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(UbicacionService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new Ubicacion(0, Departamento.MENDOZA, 'AAAAAAA', 0, 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a Ubicacion', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new Ubicacion()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a Ubicacion', () => {
        const returnedFromService = Object.assign(
          {
            departamento: 'BBBBBB',
            calle: 'BBBBBB',
            numero: 1,
            descripcion: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of Ubicacion', () => {
        const returnedFromService = Object.assign(
          {
            departamento: 'BBBBBB',
            calle: 'BBBBBB',
            numero: 1,
            descripcion: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a Ubicacion', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
